ds1077l-bus-set: Don't try to be too smart about the WC flag.
This removes the inverted handling of the boolean value. This may be
closer to the actual behavior of the flag per the docs but it adds a
translation layer that is unnecessary and will only serve to confuse in
the end I expect. Better to encourage people to read the data sheet
carefully and reward thht good behavior by giving them an accurate
representation of the register.

This also changes the parameter to make it a flag (0|1) like the
handling of values in the mux register.